+++ src/lib.rs
@@ -290,7 +290,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 "#
             .into(),
-            // TODO: Parse .git/config '[remote "origin"].url'
             repository: "https://github.com/me/my-swanky-project".into(),
             unreleased: Changes {
                 added: vec![